feat: 프록시와 연관관계 코드 추가
