+++ Geonwoo/week5/core/src/main/java/mapping/JpaMain.java
@@ -17,24 +17,29 @@ public class JpaMain {
         tx.begin();
 
         try {
-            Member member = new Member("건우", "부천시", "상오정로", "내 집");
-            em.persist(member);
+//            Team team = new Team();
+//            team.setName("team2");
+//            em.persist(team);
+//
+//            Member member = new Member();
+//            member.setName("member2");
+//            member.setTeam(team);
+//            em.persist(member);
+//
+//            em.flush();
+//            // 1차 캐시를 초기화 해야 Proxy 테스트 가능
+//            // 안 하면, 실제 객체가 조회되기 때문에
+//            em.clear();
+//
+//            Member findMember = em.find(Member.class, member.getId());
+//            System.out.println("------Proxy");
+//            // Team은 아직 프록시 상태
+//            System.out.println("team 클래스: " + findMember.getTeam().getClass());
+//            System.out.println("------Query");
+//            // 여기서 쿼리 나갈 것.
+//            System.out.println("team 이름: " + findMember.getTeam().getName());
 
-            Item item = new Item("페퍼로니 피자", 25000, 5);
-            em.persist(item);
-
-            OrderItem orderItem = new OrderItem();
-            orderItem.setCount(3);
-            orderItem.setOrderPrice(3 * item.getPrice());
-            orderItem.setItem(item);
-            em.persist(orderItem);
-
-            Order order = new Order();
-            order.setMember(member);
-            order.setOrderDate(LocalDateTime.now());
-            order.setOrderStatus(OrderStatus.SUCCESS);
-            order.addOrderItem(orderItem); // 연관관계 편의 메서드를 통해 orderItems.setOrder(this) 안 해도 됨
-            em.persist(order);
+            List<Member> members = em.createQuery("select m from Member m", Member.class).getResultList();
 
             tx.commit();
         }catch (Exception e) {
