--- Geonwoo/week5/core/src/main/java/mapping/OrderStatus.java
@@ -1,5 +1,0 @@
-package mapping;
-
-public enum OrderStatus {
-    CANCLE, SUCCESS
-}
